fb.c: Fix fb offset computations to use fb.stride
The lines in the mmapped framebuffer were wider than the width
times the pixel size.  Thus using the width to compute vertical offsets
resulted in a picture that was vertically compressed at best
(when the allocated width was a multiple of the visual width),
and completely skewed at worst.  Since the line_length parameter correctly
describes how to compute vertical offsets, using that fixed the problem

Tested on device with /dev/fb resolution of 1680x945.  (Both width
and real_width is 1680).  The pixel format is 32 bit ARGB, so using
just the with to compute the offset for the second line would give you
a value of 6720.  However, the stride is actually 6912 bytes
(corresponding to an allocated width of 1728), so if you start drawing
at offset 6720 the first 48 pixels would be invisible, and the
remaining 1632 pixels would appear skewed to the left by 48 pixels.

Tested on Zaurus poodle (qvga) and corgi/c7x0 (vga-resized)

* patch taken from zeldin fork on github and adapted after fb.c changes

Signed-off-by: Marcus Comstedt <marcus@mc.pp.se>
Signed-off-by: Andrea Adami <andrea.adami@gmail.com>