rgb.h: do not inline rgba2comp in the header (fix for gcc5)
As seen in OpenEmbeded, build with gcc5 fails:

rgb.h:55:1: warning: inline function 'rgba2comp' declared but never defined
|  rgba2comp(kx_rgba rgba, kx_ccomp *red, kx_ccomp *green,
|  ^

The fixes for C99 inline semantic are documented here:
https://gcc.gnu.org/gcc-5/porting_to.html

To follow the rest of inlined functions in kexecboot sources, do not inline
the function declaration in the header but the definition in the source.

rgba2comp is only called in fb.c and could be maybe moved there as
static inline during next code refactoring.

Signed-off-by: Andrea Adami <andrea.adami@gmail.com>