Remove machine_kernel feature.
This was meant for Distro to deploy kernels for many machines
(of the same Arch) together. It was ZUbuntu at the time.

For the Zaurus special case, with the unification of the kernels
we have to deploy one for pxa27x and one for pxa25x clamshells so
all that is pointless.
The code overhead to maintain this option seems too big.

Signed-off-by: Andrea Adami <andrea.adami@gmail.com>