kexecboot.c: Fix freeing of strings by cleaning up start_kernel()
Andrea noticed we're not freeing the strings in start_kernel(). While this
only matters in HOST_DEBUG mode where we're not kexec booting, adding new
options is still a bit hard to do. So let's the issue by allocating
load_argv and exec_argv dynamically and also any elements in them.
That way we can just free them all at the end.

Note that we are now also using an estimated size of 256 for cmdline_arg
to make things simpler.

Signed-off-by: Tony Lindgren <tony@atomide.com>